requirements.txt
fastapi
uvicorn
python-jose
passlib
scikit-learn
pandas